# chore(issues): Improve FileIOMainThreadDetector performance