FileIOMainThreadDetector is our most time consuming detector: https://sentry.sentry.io/traces/?groupBy=span.description&interval=15m&mode=aggregate&project=1&query=span.op%3Afunction%20span.description%3Arun_detector_on_data.%2A&statsPeriod=24h&visualize=%7B%22chartType%22%3A1%2C%22yAxes%22%3A%5B%22sum%28span.duration%29%22%2C%22count%28span.duration%29%22%2C%22p95%28span.duration%29%22%2C%22p99%28span.duration%29%22%5D%7D

This eliminates the need to recompile these glob patterns potentially for every transaction event and adds a synthetic benchmark to measure the (~4x) improvement:

Before:
```
$ ./bin/benchmark_detectors
1,000,000 ops
7.005 s
142,758.90 ops/s
```

After:
```
$ ./bin/benchmark_detectors
1,000,000 ops
1.806 s
553,602.94 ops/s
```