@mrduncan 2025-01-06
Great idea - we're probably reaching the limits of the accuracy of this benchmark but I do like the idea of moving to a suffix check since it's also way simpler to reason about:

Pre-compiled multiple regexes:
```
$ ./bin/benchmark_detectors
1,000,000 ops
1.782 s
561,138.73 ops/s
```

Pre-compiled single regex:
```
$ ./bin/benchmark_detectors
1,000,000 ops
1.649 s
606,374.02 ops/s
```

File path ends with:
```
$ ./bin/benchmark_detectors
1,000,000 ops
1.556 s
642,830.58 ops/s
```